backend/requirements.txt
Flask==2.2.2
flask-cors==3.1.1
transformers==4.28.1
torch==1.13.1
requests==2.28.1
Pillow==9.2.0